Ajustando gunicorn para despliegue 2

+++ scripts/stop.sh
@@ -1,2 +1,13 @@
 #!/bin/bash
-pkill -f gunicorn
+
+echo "🛑 Deteniendo Gunicorn..."
+
+# Buscar si Gunicorn está corriendo
+if pgrep -f gunicorn > /dev/null
+then
+    pkill -f gunicorn
+    echo "✅ Gunicorn detenido exitosamente."
+else
+    echo "⚠️ No se encontró ningún proceso de Gunicorn en ejecución."
+fi
+
